backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
numpy
pillow
opencv-python-headless
reportlab
supabase
python-dotenv
